refactor: TestController 테스트 코드로 이동
